tools/synthetic_dataset/requirements.txt
Faker>=24.0.0
Jinja2>=3.1.0
playwright>=1.45.0
PyMuPDF>=1.24.0
Pillow>=10.0.0
